팀 create 코드 수정

+++ k-pms/src/main/java/com/kpms/tm/service/TmServiceImpl.java
@@ -8,10 +8,12 @@ import org.springframework.stereotype.Service;
 
 import com.kpms.common.exception.APIArgsException;
 import com.kpms.common.exception.APIException;
+import com.kpms.common.util.StringUtil;
 import com.kpms.tm.dao.TmDAO;
 import com.kpms.tm.vo.TmSearchVO;
 import com.kpms.tm.vo.TmVO;
 import com.kpms.tmmbr.dao.TmMbrDAO;
+import com.kpms.tmmbr.vo.TmMbrVO;
 
 @Service
 public class TmServiceImpl implements TmService {
@@ -53,26 +55,36 @@ public class TmServiceImpl implements TmService {
 			throw new APIArgsException("400", "팀생성일이 누락되었습니다.");
 		}
 		
-		return tmDAO.createOneTm(tmVO) > 0;
+		int tmCreateCount = tmDAO.createOneTm(tmVO);
+		if (tmCreateCount > 0) {
+			List<TmMbrVO> tmMbrList = tmVO.getTmMbrList();
+			if (tmMbrList == null || tmMbrList.isEmpty()) {
+				throw new APIArgsException("404", "팀원을 선택하세요");
+			}
+			for (TmMbrVO tmMbr: tmMbrList) {
+				tmMbr.setTmId(tmVO.getTmId());
+				tmMbr.setCrtr(tmVO.getCrtr());
+				tmMbr.setMdfyr(tmVO.getMdfyr());
+				tmMbrDAO.createOneTmMbr(tmMbr);
+			}
+		}
+		return tmCreateCount > 0;
 	}
 
 	@Override
 	public boolean updateOneTm(TmVO tmVO) {
+		if (tmVO.getTmNm() == null || tmVO.getTmNm().trim().length() == 0) {
+			throw new APIArgsException("400", "팀명이 누락되었습니다.");
+		}
+		if (tmVO.getTmHdId() == null || tmVO.getTmHdId().trim().length() == 0) {
+			throw new APIArgsException("400", "팀장이 누락되었습니다.");
+		}
+		if (tmVO.getTmCrtDt() == null || tmVO.getTmCrtDt().trim().length() == 0) {
+			throw new APIArgsException("400", "팀생성일이 누락되었습니다.");
+		}
+		
 		tmMbrDAO.deleteTmMbrByTmId(tmVO.getTmId());
 		boolean result = tmDAO.updateOneTm(tmVO) > 0;
-		/*
-		 * TmVO orgnTmVO = tmDAO.readOneTmVOByTmId(tmVO.getTmId()); if (result) { String
-		 * hdTmMbrId = tmMbrDAO.readAllTmMbrVO(tmVO.getTmId()) .stream() .filter(vo ->
-		 * vo.getEmpId().equals(orgnTmVO.getTmHdId())) .map(vo -> vo.getTmMbrId())
-		 * .findFirst().orElse(null); String newTmHdId = tmVO.getTmHdId();
-		 * 
-		 * // 원래 팀장과 신규팀장이 같으면 delete 안하고 바뀌면 delete if (hdTmMbrId != null) { if
-		 * (orgnTmVO.getTmHdId().equals(newTmHdId)) {
-		 * 
-		 * } else {
-		 * 
-		 * tmMbrDAO.deleteOneTmMbrByTmMbrId(hdTmMbrId); } } }
-		 */
 		return result; 
 	}
